+++ app/User.php
@@ -22,6 +22,6 @@ class User extends Users
 
     public function owner()
     {
-        return $this->belongsTo(Owner::class);
+        return $this->belongsTo(config('enso.config.ownerModel'));
     }
 }