+++ database/factories/OwnerFactory.php
@@ -1,9 +1,8 @@
 <?php
 
-use App\Owner;
 use Faker\Generator as Faker;
 
-$factory->define(Owner::class, function (Faker $faker) {
+$factory->define(config('enso.config.ownerModel'), function (Faker $faker) {
     return [
         'name' => $faker->firstName,
         'description' => $faker->sentence,